backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.8.0
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
autopep8==2.3.2
awslambdaric==3.0.0
babel==2.16.0
backcall==0.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
billiard==4.2.1
black==25.1.0
bleach==6.2.0
blessed==1.21.0
blinker==1.9.0
boto3==1.38.12
botocore==1.38.12
Brotli==1.1.0
cachetools==5.5.0
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels==4.3.0
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmake==4.2.1
colorama==0.4.6
contourpy==1.3.2
cron-descriptor==1.4.5
cryptography==45.0.3
cssselect2==0.8.0
customtkinter==5.2.2
cvzone==1.6.1
cycler==0.12.1
darkdetect==0.8.0
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
Django==5.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-picklefield==3.3
django-q==1.3.9
django-simple-history==3.11.0
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dlib @ file:///C:/Users/hp/Desktop/FirstclickNew/backend/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
dnspython==2.6.1
docopt==0.6.2
drf-nested-routers==0.94.2
email_validator==2.2.0
emoji==2.14.1
et_xmlfile==2.0.0
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.6.0
faiss-cpu==1.10.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.21.1
filelock==3.15.4
fire==0.7.1
flasgger==0.9.7.1
flash==1.0.3
Flask==2.3.3
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Mail==0.10.0
flask-sock==0.7.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.12.19
fonttools==4.59.0
fpdf==1.7.2
frozenlist==1.8.0
fsspec==2025.7.0
future==1.0.0
gast==0.6.0
gdown==5.2.1
geoip2==5.1.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.72.0
googletrans==4.0.2
greenlet==3.2.4
groq==1.0.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.14.0
h2==3.2.0
h5py==3.13.0
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.1
httpcore==1.0.7
httplib2==0.31.0
httptools==0.6.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
hyperframe==5.2.0
idna==3.10
ipython==8.12.3
isort==6.0.1
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.2
Jinja2==3.1.6
jinxed==1.3.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kaleido==1.0.0
keras==3.14.0
kiwisolver==1.4.9
kombu==5.5.4
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.25
langchain-text-splitters==0.3.3
langsmith==0.2.3
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.5
lightphe==0.0.23
llvmlite==0.44.0
logistro==1.1.0
lxml==5.4.0
lz4==4.4.4
mangum==0.20.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.10.8
matplotlib-inline==0.1.7
maxminddb==2.8.2
mdurl==0.1.2
mediapipe==0.10.18
metatrader5==5.0.5488
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
mtcnn==1.0.0
multidict==6.7.0
mypy_extensions==1.1.0
mysql-connector-python==9.0.0
mysqlclient==2.2.7
namex==0.0.8
narwhals==2.1.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
num2words==0.5.14
numba==0.61.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
orjson==3.11.2
oscrypto==1.3.0
packaging==25.0
pandas==2.2.3
pandas-ta==0.4.71b0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20250506
pdfplumber==0.11.7
pefile==2024.8.26
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.3.8
plotly==6.3.0
prompt_toolkit==3.0.51
propcache==0.4.1
proto-plus==1.26.1
protobuf==7.34.1
psutil==7.2.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.7.0
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
pydyf==0.11.0
Pygments==2.18.0
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.9.0
PyMuPDF==1.27.1
PyMySQL==1.1.1
pyparsing==3.3.2
pypdf==5.6.0
pypdfium2==4.30.0
pyphen==0.17.2
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.9
pytz==2025.2
pywin32==309
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzk==0.9
pyzmq==26.2.1
qrcode==8.2
razorpay==2.0.0
redis==5.0.1
referencing==0.36.2
regex==2025.7.34
reportlab==4.4.1
requests==2.32.3
requests-toolbelt==1.0.0
retina-face==0.0.17
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.27.0
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.14.1
sentence-transformers==5.1.0
sentencepiece==0.2.0
setuptools==80.10.2
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.1
snapshot-restore-py==1.0.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.37.2
streamlit==1.40.2
svglib==1.5.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
tf_keras==2.21.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tkcalendar==1.6.1
tokenizers==0.21.4
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.0
typer==0.12.3
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.4.0
uvicorn==0.30.3
vanna==0.7.9
vine==5.1.0
virtualenv==20.26.3
waitress==3.0.0
watchdog==6.0.0
watchfiles==0.22.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
websockets==12.0
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.11.0
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xhtml2pdf==0.2.17
yarg==0.1.9
yarl==1.22.0
zopfli==0.2.3.post1
